+++ src/main/java/dev/germane/customer/CustomerDao.java
@@ -8,10 +8,10 @@ import java.util.Optional;
 public interface CustomerDao {
 
     List<Customer> selectAllCustomers();
-    Optional<Customer> selectCustomerById(Integer customerId);
+    Optional<Customer> selectCustomerById(Long customerId);
     void insertCustomer(Customer customer);
     boolean existsCustomerWithEmail(String email);
-    boolean existsCustomerWithId(Integer customerId);
-    void deleteById(Integer customerId);
+    boolean existsCustomerWithId(Long customerId);
+    void deleteById(Long customerId);
     void updateCustomer(Customer customer);
 }
