+++ src/main/java/dev/germane/customer/CustomerRepository.java
@@ -3,8 +3,8 @@ package dev.germane.customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CustomerRepository
-        extends JpaRepository<Customer, Integer> {
+        extends JpaRepository<Customer, Long> {
 
     boolean existsCustomerByEmail(String email);
-    boolean existsCustomerById(Integer customerId);
+    boolean existsCustomerById(Long customerId);
 }
